// packages/ui-components/src/input/enums.ts
export enum InputSize {
  RESPONSIVE = "responsive",
  NORMAL = "normal",
  SMALL = "small",
}

export enum InputColor {
  PRIMARY = "primary",
  TRANSPARENT = "transparent",
  ERROR = "error",
}

export enum InputType {
  TEXT = "text",
  EMAIL = "email",
  PASSWORD = "password",
  NUMBER = "number",
}

export enum IconPlacement {
  LEFT = "left",
  RIGHT = "right",
}